feat: Parameter to send custom page range when splitting pdf
To match the python feature: https://github.com/Unstructured-IO/unstructured-python-client/pull/125

Add a client-side param called `splitPdfPageRange` which takes a list of two integers, `[start, end]`. If `splitPdfPage` is `true` and a range is set, slice the doc from `start` up to and including `end`. Only this page range will be sent to the API. The subset of pages is still split up as needed. If `[start, end]` is out of bounds, throw an error to the user.