fix: Improve retry mechanisms for split pdf logic
HTTPClient hooks.

The autogenerated code takes these values from the `x-speakeasy-retries` section of our OpenAPI
spec. However, when we call `retry` directly like this, we need to be able to pass the values
in. Once we have a reference to the SDK in the hook context, we can make requests via
`s.general.partition` and avoid duplicating logic.