increase client concurrency to 50 (#133)
### Notes

Tested concurrency settings with a 2500 page PDF.  

Notes on various settings:

15 threads: 29 minutes
50 threads: 11 minutes
150 threads: 5 minutes


Memory usage:

15 threads:

Partition of a set of 530464 objects. Total size = 150187404 bytes.
(~150MB)

50 threads:
Partition of a set of 530606 objects. Total size = 150206646 bytes.
(~150MB)

150 threads:
Partition of a set of 530654 objects. Total size = 150221029 bytes.
(~150MB)

For now proposing we increase this to a max of 50. We can always keep
increasing later.