fix/ensure array params are sent to the server correctly (#151)
Fixes #135 

The SDK sends the wrong formdata when the user adds an array param. For
instance, the following call:
```
client.general.partition({
    partitionParameters: {
        files: {
            content: data,
            fileName: filename
        },
        extractImageBlockTypes: ["Image", "Table"],
    }
})
```

Sends the list as one comma separated FormData value:
`extract_image_block_types: "Image,Table"'.

The server will ignore this unless it's sent with multiple keys like so:
`extract_image_block_types[]: "Image", extract_image_block_types[]:
"Table"`

We addressed this before in #122 by adjusting the request body before
sending it, but this code path is for pdf splitting. A better solution
is to add a new hook that will clean up the form data for every request.
Then we can remove the modification in the splitting code.